Reflected reworked fetching methods for images and added format fetching, as to get them dynamically

+++ backend/src/repositories/ImageRepository.js
@@ -11,6 +11,7 @@ class ImageRepository {
       throw err;
     }
   }
+
   async createMultiple(files) {
     try {
       return await FileInfo.bulkCreate(files, { ignoreDuplicates: true });
@@ -18,17 +19,42 @@ class ImageRepository {
       throw err;
     }
   }
-  async getById(id) {
-    return await FileInfo.findByPk(id);
+
+  async getByQuery(query, first = 0, rows = 10) {
+    const options = {
+      where: {
+        name: { [Op.iLike]: query + "%" },
+      },
+      offset: first,
+      limit: rows,
+    };
+    return await FileInfo.findAndCountAll(options);
   }
+
+  async getById(id) {
+    const options = {
+      where: { id },
+    };
+    return await FileInfo.findOne(options);
+  }
+
+  async getAll(first = 0, rows = 10) {
+    const options = {
+      offset: first,
+      limit: rows,
+    };
+    return await FileInfo.findAndCountAll(options);
+  }
+
   /**
    * Only use in bulk operations
    */
-  async getAll() {
+  async _getAll() {
     return await FileInfo.findAll();
   }
+
   async deleteByIds(ids) {
-    if (Array.isArray(ids)) {
+    if (Array.isArray(ids) && ids.length > 0) {
       return await FileInfo.destroy({
         where: {
           id: {
